feat: Add test infrastructure and fix integration issues
- Set up xUnit test project with FluentAssertions and Moq
- Created DDD-compliant test structure (Unit/Integration folders)
- Added BusNumber value object unit tests (11 test cases)
- Documented integration tests performed on Dec 31, 2024
- Created API client for frontend-backend integration

Bug Fixes:
- Fixed BusNumber ordering in BusRepository (value object comparison)
- Fixed type conversion in BusinessInsightsController (double to decimal)
- Fixed missing using directives in DailyOperationConfiguration
- Fixed method name syntax error in FleetOptimizationService
- Configured EF Core InMemory database for testing

Test Coverage:
- Integration tests documented with detailed results
- API endpoints tested: KPIs, Fleet Status, ROI Summary
- Mock data seeding verified (20 buses, 10 routes, 7277 operations)
- Full-stack integration confirmed working

Tested on Lenovo E15 Gen3 at airport before Paris/Valencia trip

+++ backend/FleetManagement.API/Controllers/BusinessInsightsController.cs
@@ -149,7 +149,7 @@ public class BusinessInsightsController : ControllerBase
 
         var totalWasted = wastefulTrips.Sum(r => r.WastedCost);
         var annualizedWaste = totalWasted * (365m / days);
-        var potentialRevenue = overcrowdedTrips.Sum(r => (r.OccupancyPercent - 85) * 2.5m); // $2.50 per lost passenger
+        var potentialRevenue = overcrowdedTrips.Sum(r => ((decimal)r.OccupancyPercent - 85) * 2.5m); // $2.50 per lost passenger
 
         var analysis = new EmptyBusAnalysis
         {
@@ -159,7 +159,7 @@ public class BusinessInsightsController : ControllerBase
                 RouteNumber = r.RouteNumber,
                 TimeSlot = r.TimeSlot,
                 AveragePassengers = (int)r.AvgPassengers,
-                OccupancyPercent = r.OccupancyPercent,
+                OccupancyPercent = (decimal)r.OccupancyPercent,
                 TripsPerPeriod = r.TripCount,
                 WastedCost = r.WastedCost,
                 Recommendation = r.Recommendation,
@@ -170,8 +170,8 @@ public class BusinessInsightsController : ControllerBase
                 RouteNumber = r.RouteNumber,
                 TimeSlot = r.TimeSlot,
                 AveragePassengers = (int)r.AvgPassengers,
-                OccupancyPercent = r.OccupancyPercent,
-                LostRevenueEstimate = (r.OccupancyPercent - 85) * 2.5m * r.TripCount,
+                OccupancyPercent = (decimal)r.OccupancyPercent,
+                LostRevenueEstimate = ((decimal)r.OccupancyPercent - 85) * 2.5m * r.TripCount,
                 Recommendation = "Add bus to capture more passengers"
             }).ToList(),
             TotalWastedThisPeriod = totalWasted,
